Allow Sidekiq retry options in ActiveJob
https://github.com/mperham/sidekiq/commit/c685d6928ca6aec1f7569ff69871ef9a8dfb2738
has introduced an exception that is raised on an attempt to include
Sidekiq::Worker to an ActiveJob::Base descendant.

That was done in order to prevent options not supported by ActiveJob,
specifically `queue`. See https://github.com/mperham/sidekiq/issues/2424

Support for `set` was added later on, including the setting of `retry`
from ActiveJob.
https://github.com/mperham/sidekiq/commit/d6538b0b4f3c419ea969dc6710c6431b76a05d2b

This change bridges the gap by allowing ActiveJob::Base descendants to
include `Sidekiq::Options` and use `retry` option of
`sidekiq_options`, and also `sidekiq_retry_in`, and
`sidekiq_retries_exhausted`.
