+++ lib/sidekiq/worker.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "sidekiq/client"
+require "sidekiq/worker/options"
 
 module Sidekiq
   ##
@@ -24,12 +25,10 @@ module Sidekiq
     attr_accessor :jid
 
     def self.included(base)
-      raise ArgumentError, "You cannot include Sidekiq::Worker in an ActiveJob: #{base.name}" if base.ancestors.any? { |c| c.name == "ActiveJob::Base" }
+      raise ArgumentError, "You can only include Sidekiq::Worker::Options in an ActiveJob: #{base.name}" if base.ancestors.any? { |c| c.name == "ActiveJob::Base" }
 
+      base.include(Options)
       base.extend(ClassMethods)
-      base.sidekiq_class_attribute :sidekiq_options_hash
-      base.sidekiq_class_attribute :sidekiq_retry_in_block
-      base.sidekiq_class_attribute :sidekiq_retries_exhausted_block
     end
 
     def logger
@@ -71,8 +70,6 @@ module Sidekiq
     end
 
     module ClassMethods
-      ACCESSOR_MUTEX = Mutex.new
-
       def delay(*args)
         raise ArgumentError, "Do not call .delay on a Sidekiq::Worker class, call .perform_async"
       end
@@ -123,20 +120,7 @@ module Sidekiq
       # In practice, any option is allowed.  This is the main mechanism to configure the
       # options for a specific job.
       def sidekiq_options(opts = {})
-        # stringify
-        self.sidekiq_options_hash = get_sidekiq_options.merge(Hash[opts.map { |k, v| [k.to_s, v] }])
-      end
-
-      def sidekiq_retry_in(&block)
-        self.sidekiq_retry_in_block = block
-      end
-
-      def sidekiq_retries_exhausted(&block)
-        self.sidekiq_retries_exhausted_block = block
-      end
-
-      def get_sidekiq_options # :nodoc:
-        self.sidekiq_options_hash ||= Sidekiq.default_worker_options
+        super
       end
 
       def client_push(item) # :nodoc:
@@ -147,72 +131,6 @@ module Sidekiq
         end
 
         Sidekiq::Client.new(pool).push(item)
-      end
-
-      def sidekiq_class_attribute(*attrs)
-        instance_reader = true
-        instance_writer = true
-
-        attrs.each do |name|
-          synchronized_getter = "__synchronized_#{name}"
-
-          singleton_class.instance_eval do
-            undef_method(name) if method_defined?(name) || private_method_defined?(name)
-          end
-
-          define_singleton_method(synchronized_getter) { nil }
-          singleton_class.class_eval do
-            private(synchronized_getter)
-          end
-
-          define_singleton_method(name) { ACCESSOR_MUTEX.synchronize { send synchronized_getter } }
-
-          ivar = "@#{name}"
-
-          singleton_class.instance_eval do
-            m = "#{name}="
-            undef_method(m) if method_defined?(m) || private_method_defined?(m)
-          end
-          define_singleton_method("#{name}=") do |val|
-            singleton_class.class_eval do
-              ACCESSOR_MUTEX.synchronize do
-                undef_method(synchronized_getter) if method_defined?(synchronized_getter) || private_method_defined?(synchronized_getter)
-                define_method(synchronized_getter) { val }
-              end
-            end
-
-            if singleton_class?
-              class_eval do
-                undef_method(name) if method_defined?(name) || private_method_defined?(name)
-                define_method(name) do
-                  if instance_variable_defined? ivar
-                    instance_variable_get ivar
-                  else
-                    singleton_class.send name
-                  end
-                end
-              end
-            end
-            val
-          end
-
-          if instance_reader
-            undef_method(name) if method_defined?(name) || private_method_defined?(name)
-            define_method(name) do
-              if instance_variable_defined?(ivar)
-                instance_variable_get ivar
-              else
-                self.class.public_send name
-              end
-            end
-          end
-
-          if instance_writer
-            m = "#{name}="
-            undef_method(m) if method_defined?(m) || private_method_defined?(m)
-            attr_writer name
-          end
-        end
       end
     end
   end
