+++ test/test_rails.rb
@@ -5,9 +5,21 @@ require 'active_job'
 describe 'ActiveJob' do
   it 'does not allow Sidekiq::Worker in AJ::Base classes' do
     ex = assert_raises ArgumentError do
-      c = Class.new(ActiveJob::Base)
-      c.send(:include, Sidekiq::Worker)
+      Class.new(ActiveJob::Base) do
+        include Sidekiq::Worker
+      end
     end
-    assert_includes ex.message, "cannot include"
+    assert_includes ex.message, "can only include Sidekiq::Worker::Options"
+  end
+
+  it 'allows Sidekiq::Options in AJ::Base classes' do
+    Class.new(ActiveJob::Base) do
+      include Sidekiq::Worker::Options
+      sidekiq_options retry: true
+      sidekiq_retry_in { |count, _exception| count * 10 }
+      sidekiq_retries_exhausted do |msg, _exception|
+        Sidekiq.logger.warn "Failed #{msg['class']} with #{msg['args']}: #{msg['error_message']}"
+      end
+    end
   end
 end
